+++ gradle.properties
@@ -1,11 +1,11 @@
 projectVersion=1.0.0-SNAPSHOT
-projectGroup=io.micronaut.build-plugin-sourcegen
+projectGroup=io.micronaut.build.plugin.sourcegen
 
-title=Micronaut build-plugin-sourcegen
-projectDesc=TODO
+title=Micronaut Build Plugin Sourcegen
+projectDesc=Project for generating sources of build plugin tasks
 projectUrl=https://micronaut.io
 githubSlug=micronaut-projects/micronaut-build-plugin-sourcegen
-developers=Graeme Rocher
+developers=Andriy Dmytruk
 
 org.gradle.jvmargs=-Xmx2g -Dfile.encoding=UTF-8
 org.gradle.configuration-cache=false
