+++ settings.gradle
@@ -14,9 +14,9 @@ rootProject.name = 'micronaut-build-plugin-sourcegen-parent'
 include 'micronaut-build-plugin-sourcegen-bom'
 include 'micronaut-build-plugin-sourcegen-annotations'
 include 'micronaut-build-plugin-sourcegen-generator'
-include 'test-suite-common'
-include 'test-suite-gradle'
-include 'test-suite-maven'
+include 'test-suite-common-java'
+include 'test-suite-gradle-java'
+include 'test-suite-maven-java'
 
 enableFeaturePreview 'TYPESAFE_PROJECT_ACCESSORS'
