+++ src/main/docs/guide/gettingStarted/mavenModule.adoc
@@ -1,0 +1,32 @@
+The Maven module should have the same dependencies and also the dependency on the common module.
+
+dependency:micronaut-build-plugin-sourcegen-annotations[groupId="io.micronaut.build.plugin.sourcegen"]
+
+dependency:test-plugin-common[scope="compileOnly", groupId="io.micronaut.test"]
+
+dependency:micronaut-build-plugin-sourcegen-generator[scope="annotationProcessor", groupId="io.micronaut.build.plugin.sourcegen"]
+
+dependency:test-plugin-common[scope="annotationProcessor", groupId="io.micronaut.test"]
+
+NOTE: Adding the common plugin to the annotation processor paths is currently required to retrieve javadoc.
+
+Use the link:{api}/io/micronaut/sourcegen/annotations/GenerateMavenMojo.html[GenerateMavenMojo] annotation to trigger generation of Maven Plugin sources.
+
+snippet::io.micronaut.sourcegen.example.plugin.maven.GenerateMojoTrigger[project-base="test-suite-maven", tags="content", source="main"]
+
+<1> Trigger generation of a mojo. Specify the task from common module annotated with link:{api}/io/micronaut/sourcegen/annotations/PluginTask.html[PluginTask] as source. Based on the prefix, `AbstractGenerateSimpleRecordMojo` will be generated.
+<2> If you create another task, you can generate another Mojo for it.
+
+Only `AbstractGenerateSimpleRecordMojo` class will be generated. The mojo will have all the specified task parameters and will call the defined task as its action.
+
+See documentation for link:{api}/io/micronaut/sourcegen/annotations/GenerateMavenMojo.html[GenerateMavenMojo] to view all the configurable properties.
+
+=== Mojo Customization
+
+Extend the Mojo to add custom Maven-specific behavior:
+
+snippet::io.micronaut.sourcegen.example.plugin.maven.GenerateSimpleRecordMojo[project-base="test-suite-maven", tags="content", source="main"]
+
+<2> Specify a name for Mojo.
+<2> Add a property for enabling and disabling the mojo.
+<3> Add the generated folder to sources.
