--- src/main/docs/guide/quickStart.adoc
@@ -1,1 +1,0 @@
-TODO
