+++ test-suite-common-java/build.gradle.kts
