+++ test-suite-common-java/src/main/java/io/micronaut/sourcegen/example/plugin/GenerateSimpleResourceTask.java
