+++ test-suite-gradle-java/src/main/resources/META-INF/gradle-plugins/io.micronaut.sourcegen.test.properties
