+++ test-suite-gradle-java/src/test/java/io/micronaut/sourcegen/example/plugin/gradle/TestPluginTest.java
