+++ test-suite-maven-java/src/main/java/io/micronaut/sourcegen/example/plugin/maven/GenerateSimpleResourceMojo.java
