+++ test-suite-maven-java/src/main/resources/META-INF/maven/plugin.xml
