+++ test-suite-maven-java/src/main/resources/META-INF/plexus/components.xml
