+++ test-suite-maven-java/src/test/java/io/micronaut/sourcegen/example/plugin/maven/AbstractMavenPluginTest.java
