+++ test-suite-maven-java/src/test/resources/test-pom.xml
