+++ test-suite-maven-java/src/test/resources/test-resource-pom.xml
