+++ test-suite-maven-java/src/test/resources/test-resource-repeat-pom.xml
